CSS additions
Trying to add CSS to make DrumMachineD more colorful, and also increased the number of repetitions to allow the user to jam more without having to constantly press play. Future plans include adding functionality to the stop button, adding a button to enable infinite playback for the track, and improve CSS for the buttons and dropdown menus.